update getMarketplaces to proposed new structure
