+++ lib/helpers/getMarketplaces.js
@@ -1,4 +1,3 @@
-const { parseEndpoint } = require('../parseEndpoint');
 const parseMarketplaceData = require('../parsers/marketplaceData');
 
 // TODO: upgrade to call ListMarketplaceParticipationsByNextToken when a NextToken
@@ -22,19 +21,6 @@ const parseMarketplaceData = require('../parsers/marketplaceData');
  * @param {boolean} hasSellerSuspendedListings - true if there are seller suspended listings in this account on this market
  */
 
-/**
- * Call MWS ListMarketplaceParticipations, return parsed results
- * @example
- * const marketplaces = (async () => await mws.getMarketplaces())();
- * (async function() {
- *    const result = await mws.getMarketplaces();
- *    console.log(result);
- * })();
- * @return {MarketDetail}
- */
+const getMarketplaces = api => api.parseEndpoint(parseMarketplaceData)('ListMarketplaceParticipations');
 
-const getMarketplaces = parseEndpoint(parseMarketplaceData)('ListMarketplaceParticipations');
-
-module.exports = {
-    getMarketplaces,
-};
+module.exports = getMarketplaces;
